moved to seperate files and fixed game speed
