+++ old_tetris.py
@@ -1,94 +1,32 @@
-import pygame
-
 import figure
 
 
-class Tetris(object):
-    pressing_down = False
+# Logic of the tetris program
+class Tetris:
     level = 2
     score = 0
+    state = ""
     field = []
     height = 0
     width = 0
-    time_elapsed = 0
     x = 100
     y = 60
     zoom = 20
     figure = None
-    """
-    A single instance of this class is responsible for
-    managing which individual game state is active
-    and keeping it updated. It also handles many of
-    pygame's nuts and bolts (managing the event
-    queue, framerate, updating the display, etc.).
-    and its run method serves as the "game loop".
-    """
 
-    def __init__(self, width, height, screen, statess, start_state):
-        """
-        Initialize the Game object.
-
-        screen: the pygame display surface
-        states: a dict mapping state-names to GameState objects
-        start_state: name of the first active game state
-        """
+    # initialize the tetris object
+    def __init__(self, height, width):
         self.done = False
-        self.screen = screen
-        self.width = width
         self.height = height
+        self.width = width
         self.field = []
-        self.clock = pygame.time.Clock()
-        self.fps = 25
         self.score = 0
-        self.states = statess
-        self.state_name = start_state
-        self.state = self.states[self.state_name]
         for i in range(height):
             new_line = []
             for j in range(width):
                 new_line.append(0)
             self.field.append(new_line)
 
-    def event_loop(self):
-        """Events are passed for handling to the current state."""
-        for event in pygame.event.get():
-            self.state.get_event(event)
-
-    def flip_state(self):
-        """Switch to the next game state."""
-        current_state = self.state_name
-        next_state = self.state.next_state
-        self.state.done = False
-        self.state_name = next_state
-        persistent = self.state.persist
-        self.state = self.states[self.state_name]
-        self.state.startup(persistent)
-
-    def run(self):
-        """
-        Pretty much the entirety of the game's runtime will be
-        spent inside this while loop.
-        """
-        while not self.done:
-            if self.figure is None:
-                self.new_figure()
-            dt = self.clock.tick(self.fps)
-            self.event_loop()
-            self.update(dt)
-            self.draw()
-            pygame.display.update()
-
-    def update(self, dt):
-        """
-        Check for state flip and update active state.
-
-        dt: milliseconds since last frame
-        """
-        if self.state.quit:
-            self.done = True
-        elif self.state.done:
-            self.flip_state()
-        self.state.update(dt)
 
     def draw(self):
         """Pass display surface to active state for drawing."""
@@ -116,7 +54,7 @@ class Tetris(object):
         lines = 0
         for i in range(1, self.height):
             zeros = 0
-            for j in range(self.height):
+            for j in range(self.width):
                 if self.field[i][j] == 0:
                     zeros += 1
             if zeros == 0:
@@ -164,3 +102,4 @@ class Tetris(object):
         self.figure.rotate()
         if self.collision():
             self.figure.rotation = old_rotation
+
